refactor: move ContextSchema validation to serialize
